+++ src/enums/dependencies.py
@@ -1,17 +1,21 @@
 from enum import auto as auto_enum
+
 from typing_extensions import override
+
 from src.enums import CaseInsensitiveEnum
 
 
 class Dependencies(CaseInsensitiveEnum):
     FFMPEG = auto_enum()
     FRAME_FORGE = auto_enum()
+    MKBRR = auto_enum()
 
     @override
     def __str__(self) -> str:
         dep_map = {
             Dependencies.FFMPEG: "FFMPEG",
             Dependencies.FRAME_FORGE: "FrameForge",
+            Dependencies.MKBRR: "mkbrr",
         }
         return dep_map[self]
 
@@ -26,6 +30,11 @@ class Dependencies(CaseInsensitiveEnum):
                 "cfg_var": "frame_forge",
                 "app_folder": "frame_forge",
                 "executable": "frameforge",
+            },
+            Dependencies.MKBRR: {
+                "cfg_var": "mkbrr",
+                "app_folder": "mkbrr",
+                "executable": "mkbrr",
             },
         }
         return dep_map[self]